requirements.txt
jupyter~=1.1.1
pip~=24.2
attrs~=23.2.0
wheel~=0.44.0
pillow~=10.3.0
tornado~=6.4.1
Jinja2~=3.1.4
filelock~=3.15.4
numpy~=1.26.4
torch~=2.7.0
transformers~=4.51.3
huggingface_hub==0.31.1
torchvision==0.20